Fix nvToolsExt.h handling of NVTX_EXPORT_API to not require user to manually define NVTX_DECLSPEC
Previously, defining NVTX_EXPORT_API wasn't sufficient to make the NVTX v3
headers buildable as a dynamic library with the API exported -- the user also
had to define NVTX_DECLSPEC as an appropriate prefix to cause the symbol to be
exported, which is different on each platform, and non-obvious.  This change
allows users to still specify NVTX_DECLSPEC as an override, so the old usage
is still supported, but now defining NVTX_EXPORT_API without also defining
NVTX_DECLSPEC results in NVTX_DECLSPEC being defined appropriately to export
symbols on Windows and Linux.

Tested on Windows and Linux, with x64, x86, ARM64, and ARM64EC.  Confirmed
that using NVTX_SET_NAME_MANGLING_OPTIONS in a function ensures it will not
have any name mangling or decorations applied to the exported symbol.  Add
use of NVTX_SET_NAME_MANGLING_OPTIONS in all NVTX entry points in a follow
up commit.

Also write explanations of these macros in the comments and fix some other
nearby comments.