NVTX extension APIs return 0 if no tool attached
instead of `-1`. The main reason for this change is the
`nvtxDomainIsEnabled` function, which should either return `1` when
the domain is enabled and a tool attached or `0`.

Push/pop APIs are an exception and return `NVTX_NO_PUSH_POP_TRACKING`
(`-2`) when no tool is attached or the tool doesn't track push/pop
levels.